paling latest 18/12/23

+++ app/Http/Livewire/ViewBloodTest/ViewBloodTestPage.php
@@ -26,6 +26,7 @@ class ViewBloodTestPage extends Component
     public $calcium;
     public $blood_status;
     public $BloodTest;
+    // public $blood_availability;
 
     public function openModalViewTestMore($id){
         $this->viewTestMoreModal = true;
@@ -85,37 +86,74 @@ class ViewBloodTestPage extends Component
         ]);
     }
 
+    // public function delete($id){
+
+    //     $BloodTest = BloodInformation::find($id);
+    //     $BloodTest->delete();
+    //     $this->test_id = $BloodTest->test_id;
+    //     $this->blood_id = $BloodTest->blood_id;
+    //     $this->blood_group = $BloodTest->blood_group;
+    //     $this->date = $BloodTest->date;
+    //     // $this->blood_availability = $BloodTest->blood_availability;
+    //     $this->white_blood_cells = $BloodTest->white_blood_cells;
+    //     $this->red_blood_cells = $BloodTest->red_blood_cells;
+    //     $this->haemoglobin = $BloodTest->haemoglobin;
+    //     $this->hematrocrit = $BloodTest->hematrocrit;
+    //     $this->platelets = $BloodTest->platelets;
+    //     $this->hiv_test = $BloodTest->hiv_test;
+    //     $this->sodium = $BloodTest->sodium;
+    //     $this->potassium = $BloodTest->potassium;
+    //     $this->calcium = $BloodTest->calcium;
+    //     $this->blood_status = $BloodTest->blood_status;
+
+    //     $this->dialog([
+    //         'title'       => 'Data Deleted!',
+    //         'description' => 'Data was successfully deleted',
+    //         'icon'        => 'success'
+    //     ]);
+
+    //     $bloodStock = BloodStock::where('blood_type' ,$this->blood_group )->first();
+
+    //     $updateBloodStock = BloodStock::where('blood_type' ,$this->blood_group )
+    //             ->update(['quantity' => $bloodStock->quantity - 1]);
+
+    // }
+
     public function delete($id){
 
         $BloodTest = BloodInformation::find($id);
-        $BloodTest->delete();
-        $this->test_id = $BloodTest->test_id;
-        $this->blood_id = $BloodTest->blood_id;
-        $this->blood_group = $BloodTest->blood_group;
-        $this->date = $BloodTest->date;
-        $this->white_blood_cells = $BloodTest->white_blood_cells;
-        $this->red_blood_cells = $BloodTest->red_blood_cells;
-        $this->haemoglobin = $BloodTest->haemoglobin;
-        $this->hematrocrit = $BloodTest->hematrocrit;
-        $this->platelets = $BloodTest->platelets;
-        $this->hiv_test = $BloodTest->hiv_test;
-        $this->sodium = $BloodTest->sodium;
-        $this->potassium = $BloodTest->potassium;
-        $this->calcium = $BloodTest->calcium;
-        $this->blood_status = $BloodTest->blood_status;
+        $bloodGroup = $BloodTest->blood_group;
 
-        $this->dialog([
-            'title'       => 'Data Deleted!',
-            'description' => 'Data was successfully deleted',
-            'icon'        => 'success'
-        ]);
+        if($BloodTest){
+            $BloodTest->delete();
 
-        $bloodStock = BloodStock::where('blood_type' ,$this->blood_group )->first();
+            $this->test_id = $BloodTest->test_id;
+            $this->blood_id = $BloodTest->blood_id;
+            // Assign other properties as needed...
 
-        $updateBloodStock = BloodStock::where('blood_type' ,$this->blood_group )
-                ->update(['quantity' => $bloodStock->quantity - 1]);
+            $this->dialog([
+                'title'       => 'Data Deleted!',
+                'description' => 'Data was successfully deleted',
+                'icon'        => 'success'
+            ]);
 
+            $bloodStock = BloodStock::where('blood_type', $bloodGroup)->first();
+
+            if ($bloodStock && $bloodStock->quantity > 0) {
+                $newQuantity = max(0, $bloodStock->quantity - 1);
+
+                $updateBloodStock = BloodStock::where('blood_type', $bloodGroup)
+                    ->update(['quantity' => $newQuantity]);
+            } else {
+                // Handle the case where the blood stock doesn't exist or is already at 0
+                // You might want to add error handling or other logic here
+            }
+        } else {
+            // Handle the case where BloodTest with the given ID doesn't exist
+            // You might want to add error handling or other logic here
+        }
     }
+
 
     public function render()
     {
